[101].Remove obsolete files light.svg and sprite.svg and add new SVG file eye-20-20.svg
light.svg and sprite.svg were removed because they are no longer used within the codebase.
A new SVG file has been added, which adds the image for the 'view' button.

Additional changes include:
- modify buildWebpackConfig.ts to fix publicPath in webpack.config and correctly resolve relative paths
- add 'articles' and 'article_details' endpoints to the routeConfig and added corresponding translations for 'article' namespace
- create new ArticlesPage and ArticleDetailsPage including their async versions for lazy rendering
- add 'articles' path with corresponding ArticlesIcon to the sidebar items
- add new 'eye-20-20.svg' file which will represent the 'view' button in the site

These changes improve the structure of the codebase and have prepared the project for the implementation of the articles viewing feature.